Ripple, OnClick, href and error checking
Add ripple support to menu items in the form of a manual enabling as a method or auto init in the attacher class.
Add HREF property of menu items so that a custom URL can be specified.
Add OnClick property so that custom onclick methods can be added to the menu item.
Bug fixed the remove menu item method of the nav drawer which was checking for the old identifier.
Add value checking to the constructor method of the classes to be more stable with bad implementations.